requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
-e .
    # via -r requirements.in
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   google-genai
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
attrs==25.3.0
    # via
    #   cyclopts
    #   jsonschema
    #   referencing
authlib==1.6.1
    # via fastmcp
blinker==1.9.0
    # via streamlit
cachetools==5.5.2
    # via
    #   google-auth
    #   streamlit
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   streamlit
    #   uvicorn
cryptography==45.0.6
    # via authlib
cyclopts==3.22.5
    # via fastmcp
dnspython==2.7.0
    # via email-validator
docstring-parser==0.17.0
    # via cyclopts
docutils==0.22
    # via rich-rst
email-validator==2.2.0
    # via pydantic
exceptiongroup==1.3.0
    # via fastmcp
fastmcp==2.11.2
    # via pinionai
geographiclib==2.0
    # via geopy
geopy==2.4.1
    # via pinionai-chat
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
google-api-core==2.25.1
    # via
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
    #   google-genai
google-cloud-appengine-logging==1.6.2
    # via google-cloud-logging
google-cloud-audit-log==0.3.2
    # via google-cloud-logging
google-cloud-core==2.4.3
    # via google-cloud-logging
google-cloud-logging==3.12.1
    # via pinionai-chat
google-genai==1.35.0
    # via pinionai
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.2
    # via google-cloud-logging
grpcio==1.74.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   pinionai
grpcio-status==1.74.0
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   fastmcp
    #   google-genai
    #   mcp
    #   pinionai
httpx-sse==0.4.1
    # via mcp
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
isodate==0.7.2
    # via openapi-core
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonpath-ng==1.7.0
    # via pinionai
jsonschema==4.25.0
    # via
    #   altair
    #   mcp
    #   openapi-core
    #   openapi-schema-validator
    #   openapi-spec-validator
jsonschema-path==0.3.4
    # via
    #   openapi-core
    #   openapi-spec-validator
jsonschema-specifications==2025.4.1
    # via
    #   jsonschema
    #   openapi-schema-validator
lazy-object-proxy==1.11.0
    # via openapi-spec-validator
levenshtein==0.26.1
    # via python-levenshtein
markdown==3.5.2
    # via pinionai
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
mcp==1.12.4
    # via fastmcp
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.7.0
    # via openapi-core
narwhals==2.0.1
    # via altair
numpy==2.3.2
    # via
    #   pandas
    #   pydeck
    #   streamlit
openapi-core==0.19.5
    # via fastmcp
openapi-pydantic==0.5.1
    # via fastmcp
openapi-schema-validator==0.6.3
    # via
    #   openapi-core
    #   openapi-spec-validator
openapi-spec-validator==0.7.2
    # via openapi-core
opentelemetry-api==1.36.0
    # via google-cloud-logging
packaging==25.0
    # via
    #   altair
    #   streamlit
pandas==2.3.1
    # via
    #   pinionai
    #   streamlit
parse==1.20.2
    # via openapi-core
pathable==0.4.4
    # via jsonschema-path
pillow==11.3.0
    # via
    #   pinionai-chat
    #   streamlit
pinionai==0.1.0
    # via pinionai-chat
ply==3.11
    # via jsonpath-ng
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-logging
protobuf==6.32.0
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-audit-log
    #   google-cloud-logging
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   pinionai
    #   proto-plus
    #   streamlit
pyarrow==21.0.0
    # via streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   fastmcp
    #   google-genai
    #   mcp
    #   openapi-pydantic
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via mcp
pydeck==0.9.1
    # via streamlit
pygments==2.19.2
    # via rich
pyperclip==1.9.0
    # via fastmcp
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via
    #   fastmcp
    #   pydantic-settings
python-levenshtein==0.26.1
    # via pinionai
python-multipart==0.0.20
    # via mcp
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via jsonschema-path
rapidfuzz==3.13.0
    # via levenshtein
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   google-api-core
    #   google-genai
    #   jsonschema-path
    #   streamlit
rfc3339-validator==0.1.4
    # via openapi-schema-validator
rich==14.1.0
    # via
    #   cyclopts
    #   fastmcp
    #   rich-rst
rich-rst==1.3.1
    # via cyclopts
rpds-py==0.27.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via
    #   python-dateutil
    #   rfc3339-validator
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
sse-starlette==3.0.2
    # via mcp
starlette==0.47.2
    # via mcp
streamlit==1.49.1
    # via pinionai-chat
tenacity==9.1.2
    # via
    #   google-genai
    #   streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.2
    # via streamlit
typing-extensions==4.14.1
    # via
    #   altair
    #   google-genai
    #   openapi-core
    #   opentelemetry-api
    #   pydantic
    #   pydantic-core
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via mcp
websockets==15.0.1
    # via
    #   google-genai
    #   pinionai
werkzeug==3.1.1
    # via openapi-core
xmltodict==0.14.2
    # via pinionai
zipp==3.23.0
    # via importlib-metadata
